Gradle: Improve the detection of unresolvable configurations
Align the logic to detect unresolvable configurations with the logic
used by Gradle [1]. This prevents ORT from trying to resolve some
configurations that shall not be resolved, for example the "default"
configuration in many projects, and thereby avoids issues during the
resolution which are impossible to resolve. For example, in such
configurations Gradle can sometimes not determine the correct variant of
a dependency.

This change increases the minimum supported Gradle version to 6.0 from
November 2019, which is the first release that contains the
`DeprecatableConfiguration` class. Therefore we can now also safely
assume that the `Configuration` has the `isCanBeResolved()` function,
which is available since Gradle 3.3.

[1]: https://github.com/gradle/gradle/blob/0109a3c028c6ac8c6040c5dfcbdec2e5a9c51539/subprojects/diagnostics/src/main/java/org/gradle/api/tasks/diagnostics/internal/ConfigurationDetails.java#L40-L43

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>